Rename: change components name to fit convention in picture
- change components name
- In directory, picture/[component].tsx
- In import, Picture[component]
