+++ app/ui/picture/EnlargedPreview.tsx
